document VideoDisplayGenerator

+++ src/ophys_etl/modules/segmentation/qc_utils/video_display_generator.py
@@ -1,15 +1,34 @@
 from pathlib import Path
 import tempfile
 
+from ophys_etl.modules.segmentation.qc_utils import video_utils
+
 
 class VideoDisplayGenerator(object):
     """
+    This is a helper class designed to serve up thumbnail videos
+    in a way that they can be displayed in Jupyter notebooks.
+
     In order to display videos without embedding them, the videos
     have to be in a path relative to the notebook's current working
     directory. Rather than actually write our thumbnails there,
     we will create a separate temp dir with symlinks to the actual
     thumbnails. This class will handle that process. It will also
     clean up the symlinks behind itself on deletion.
+
+    The constructor for this class takes no arguments. It just
+    creates a temporary directory that is a sub-directory of
+    the notebook's working directory. The only method any user
+    should care about is
+
+    my_video_display_generator.display_video()
+
+    which returns a dict of parameters suitable to be passed to
+    IPython.display.Video, i.e.
+
+    Video(**my_video_display_generator.display_video(thumbnail))
+
+    will cause a thumbnail video to be displayed in the notebook.
     """
     def __init__(self):
         self.this_dir = Path('.').absolute()
@@ -30,9 +49,9 @@ class VideoDisplayGenerator(object):
 
     def clean_tmp(self):
         """
-        Scan through 'tmp', cleaning out empty thumbnail dirs that have
-        accumulated while running this notebook. We need to do this here
-        because, as this class is used, .nfs files get placed in the
+        Scan through self.tmp_dir, cleaning out empty thumbnail dirs that
+        have accumulated while running this notebook. We need to do this
+        here because, as this class is used, .nfs files get placed in the
         temp dir. These are still in use when the destructor from this
         class gets called, making it impossible for us to clean up tmp
         upon deconstruction of this class.
@@ -45,24 +64,43 @@ class VideoDisplayGenerator(object):
                 continue
             if dirname.is_dir():
                 sub_contents = [fname for fname in dirname.iterdir()]
-            if len(sub_contents) == 0:
-                dirname.rmdir()
+                if len(sub_contents) == 0:
+                    dirname.rmdir()
 
     def __del__(self):
         """
         Automatically delete all of the symlinks that were written.
         """
         for f_path in self.files_written:
-            f_path.unlink()
+            if f_path.exists():
+                f_path.unlink()
 
-    def display_video(self, thumbnail, width=512, height=512):
+    def display_video(self,
+                      thumbnail: video_utils.ThumbnailVideo,
+                      width: int = 512,
+                      height: int = 512):
         """
         Display a video in this notebook. As a part of displaying the video,
         a symlink to the actual video is created somewhere under the directory
         of this notebook.
 
-        thumbnail is a ThumbnailVideo as defined in
-        segmentation/qc_utils/video_utils
+        Parameters
+        ----------
+        thumbnail: video_utils.ThumbnailVideo
+
+        width: int
+            The width of the video as you want it to appear in the notebook
+            (default = 512)
+
+        height: int
+            The height of the video as you want it to appear in the notebook
+
+        Returns
+        -------
+        params: dict
+            The parameters that need to be passed as kwargs to
+            IPython.display.Video to get the video to appear in the
+            notebook
         """
         # in case another instance of this notebook accidentally
         # deleted this notebook's tmpdir
